requirements.txt
arrow==0.12.1
certifi==2018.11.29
chardet==3.0.4
gspread==0.6.2
httplib2==0.12.0
idna==2.8
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.2
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
twilio==6.22.0
urllib3==1.24.1
